Combine InstallSchema, AvailableTables
The main issue here is that `schema.tpl` is getting the wrong data. It was getting the DB name

```php
generateCreateSql($model->srcPath, $model->db['database'], $model->tables)
```

But it should actually the full, parsed schema specification:

```
generateCreateSql($model->srcPath, $spec->database, $spec->tables)
```

So basically...  `InstallSchema` needs access to the full `$specification`
(which was only available in `AvailableTables`).  And I don't currently see
a use-case that requires exposing the schema as part of the installer model,
so I think it's simpler to merge `AvailableTables` into `InstallSchema`.